requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.59.3
deepl==1.20.0
python-multipart==0.0.20
python-dotenv==1.0.1
